requirements.txt
openai==1.12.0
manim==0.19.0
python-dotenv==1.0.1